refactor: mover responses para /schemas
